[INFRA-6259] feat: implement reproducible zip (#10)f
Alters the zip method to reset all file permissions and timestamps of the files to be zipped, prior to zipping.

The file modification time respects $SOURCE_DATE_EPOC (according to the reproducible builds spec[1]), and defaults to 1/1/1980. The file permissions are also updated to 644, to enforce standardization.

More information on reproducible builds for zip files can be found here[2].

[1]: https://reproducible-builds.org/docs/timestamps/
[2]: https://reproducible-builds.org/docs/archives/